% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{adjust.color}
\alias{adjust.color}
\title{Darken or lighten colors.}
\usage{
adjust.color(col, f)
}
\arguments{
\item{col}{Character vector with valid R colors that \code{col2rgb()} accepts.}

\item{f}{Factor controlling strength of color change. A value of 1.0 does not
change the colors, values < 1.0 make them darker, values > 1.0 ligther.}
}
\value{
A character vector with altered colors of the same length as \code{col}.
}
\description{
Darken or lighten colors.
}
