% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{classError}
\alias{classError}
\title{Classification error.}
\usage{
classError(a, b, map = NULL, exhaustive = FALSE)
}
\arguments{
\item{a}{A numeric, charactor of factor with class labels.}

\item{b}{A numeric, charactor of factor with class labels (same length as \code{a}).}

\item{map}{A mapping from labels in \code{a} to labels in \code{b}, for example
returned by \code{mapLabelsGreedy}. If \code{map} is \code{NULL},
\code{mapLabelsExhaustive} (if \code{exhaustive} is \code{TRUE}) or
\code{mapLabelsGreedy} is called to create such a mapping.}

\item{exhaustive}{If \code{TRUE} and \code{map} is \code{NULL}, use \code{mapLabelsExhaustive}
to get \code{map}. If \code{FALSE} and \code{map} is \code{NULL}, use \code{mapLabelsGreedy}.
Otherwise, \code{exhaustive} is ignored.}
}
\value{
A numerical value corresponding to the classification error.
}
\description{
Calculate the error of a given classification relative to a second one.
}
\details{
The classification error is calculated as the fraction of elements, for which
the label in \code{a} is not equal to the label given in \code{b}. Optionally,
a mapping between the two label sets can be specificed using \code{map}.
}
\seealso{
\code{\link{mapLabelsExhaustive}} and \code{\link{mapLabelsGreedy}} for
    calculation of a mapping between \code{a} and \code{b} labels.
}
