% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{edge.weights}
\alias{edge.weights}
\title{Edge reweighting to group nodes by community during plotting}
\usage{
edge.weights(community, network, weight.within = 10, weight.between = 1)
}
\arguments{
\item{community}{A community object returned by any community detection
function of \code{igraph}}

\item{network}{A graph object of class \code{igraph}}

\item{weight.within}{A scalar specifying the weight for edges connecting nodes
within the same community}

\item{weight.between}{A scalar specifying the weight for edges connecting nodes
between different communities (communities crossing edges)}
}
\value{
A numeric vector of length \code{length(E(network))} with the new
edge weights
}
\description{
Edges crossing communites are downweighted while edges within 
a community are unaffected (or up-weighted). As a result during plotting
with an \code{igraph} layout algorithm that takes into account edge
weights (such as \code{layout_with_fr}) communities are brought 
together and stand out better in the plot. This operation distorts
the "true node distances" to improve community visibility
}
