% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Emulate ggplot2 color palette.}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{numeric(1) specifying the number of colors to return.}
}
\value{
A character verctor with \code{n} colors spreading the whole rainbow.
}
\description{
Emulate ggplot2 color palette.
}
