% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griph.R
\docType{package}
\name{griph-package}
\alias{griph}
\alias{griph-package}
\title{Graph Inference of Population Heterogeneity.}
\description{
griph identifies cell types in single cell RNA-seq datasets,
allowing to control for unwanted sources of variance (e.g. cell cycle)
that may confound the cell types.
}
\details{
The most important function that combines all the steps of cell type identification
from \pkg{griph} is \code{\link{griph_cluster}}. Otherwise refer to the vignette
to see what else is available and how to optimize results for your particular
dataset.
}
\author{
\strong{Maintainer}: Panagiotis Papasaikas \email{panagiotis.papasaikas@fmi.ch}

Authors:
\itemize{
  \item Michael Stadler \email{michael.stadler@fmi.ch}
  \item Atul Sethi \email{atul.sethi@fmi.ch}
}

}
