% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griph_cluster.R
\name{griph_cluster}
\alias{griph_cluster}
\title{Wrapper function for SC_cluster}
\usage{
griph_cluster(DM, SamplingSize = 750, ref.iter = 1, use.par = FALSE,
  ncores = "all", filter = FALSE, rho = 0.25, batch.penalty = 0.5,
  ClassAssignment = rep(1, ncol(DM)), BatchAssignment = NULL, ncom = NULL,
  plotG = TRUE, maxG = 2500, fsuffix = RandString(),
  image.format = "png")
}
\arguments{
\item{DM}{Count data (n genes-by-k cells) or directly a correlation k-by-k matrix.
Required argument.}

\item{SamplingSize}{Number of sampled cells in initialization step.}

\item{ref.iter}{Number of clustering refinement iterations.}

\item{use.par}{If \code{TRUE}, use parallel versions of distance calculation
functions based on \code{\link[foreach]{foreach}} (see details).}

\item{ncores}{a numeric(1) or character(1), either specifying the number of
parallel jobs to use, or \code{"all"} (default) to use up to 90 percent of
the available cores. Ignored if \code{use.par==FALSE}.}

\item{filter}{T/F Filter genes according to cv=f( mean ) fitting. Default: \code{TRUE}.}

\item{rho}{Inverse covariance matrix regularization (graph sparsification) parameter -> [0,1].
Default=0.25.  The parameter is then automatically scaled to account for
number of variables or converted into a matrix and adjusted according to
the batch.penalty factor to account for BatchAssignment (if given).}

\item{batch.penalty}{[0,1] rho scaling factor for enforcing topological constraints
variables according to \code{BatchAssignment}. For penalty p  -> rho_same_batch=rho^(1-p),
rho_diff_batch=rho^(1+p). It is ignored if \code{BatchAssignment==NULL}.}

\item{ClassAssignment}{If available a numeric vector of length \code{k} with numeric
class labels (e.g-> c(1,2,1,1,1,2,3,3,3,1,2))}

\item{BatchAssignment}{If available a numeric vector of length \code{k} with numeric
batch labels (e.g-> c(1,2,1,1,1,2,3,3,3,1,2))}

\item{ncom}{Forces detection of a fixed number of clusters (communities).}

\item{plotG}{if \code{TRUE} plots the resulting graph}

\item{maxG}{Approximate maximal number of vertices to include when plotting the graph.}

\item{fsuffix}{A suffix added to the file names of output plots. If not given
it will use a random 5 character string.}

\item{image.format}{Specifies the format of the created images. Currently only pdf and png filetypes are supported.}
}
\value{
Currently a list with the clustering results.
}
\description{
\code{griph_cluster} takes a gene-by-cell matrix with read counts as
    input and groups individual cells into clusters of similar cell (cell types).
    The number of cell types is automatically determined.
}
\details{
TODO
}
