% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griph_order.R
\name{griph_order}
\alias{griph_order}
\title{Pseudo time ordering of cell groups}
\usage{
griph_order(gr, memb = NULL)
}
\arguments{
\item{gr}{A \code{griph} result, as returned by \code{\link{griph_cluster}}.}

\item{memb}{A membership vector of length equal to the number of cells present in gr.}
}
\description{
Calculates a pseudotime ordering of groups of cells using the affinity matrix  
    and membership assignemnt returned from \code{\link{griph_cluster}} or the affinity matrix 
    and a custom membership assignment.
}
