% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HellingerMat}
\alias{HellingerMat}
\title{Hellinger distance between columns of a matrix.}
\usage{
HellingerMat(A)
}
\arguments{
\item{A}{Numeric matrix.}
}
\value{
A square matrix of dimensions \code{ncol(A})-by-\code{ncol(A)}.
}
\description{
\code{HellingerMat} returns a matrix of Hellinger distances between
the columns of the matrix passed as argument.
}
\details{
This function calculates the Hellinger distances between columns of the matrix
given as argument. It is implemented in C++ for efficiency. For an argument
matrix with r rows and c columns, the return value is an c-by-c matrix with
all pairwise distances.
}
\examples{
x <- matrix(1:12, nrow=3, ncol=4)
HellingerMat(x)

}
