% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{keep.mknn}
\alias{keep.mknn}
\title{Keep k mutual nearest neighbors of variables given a sparse similarity matrix.}
\usage{
keep.mknn(S, k = round(sqrt(nrow(S))))
}
\arguments{
\item{S}{a sparse similarity matrix (square, symmetric, non-negative) of class CsparseMatrix}

\item{k}{integer giving the number of mutual nearest meighbors
Defaults to the rounded square root of the number of variables.}
}
\value{
a sparse Matrix of class CsparseMatrix with at most \code{length(S@x)} non-zero elements.
}
\description{
Keep k mutual nearest neighbors of variables given a sparse similarity matrix.
}
