% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{keep.mknn2}
\alias{keep.mknn2}
\title{Keep k mutual nearest neighbors of variables given a sparse similarity matrix.  
By default 10 nearest neighbors of each node are retained to avoid node isolation}
\usage{
keep.mknn2(S, k = round(sqrt(nrow(S))), quant = 0.8)
}
\arguments{
\item{S}{a sparse similarity matrix (square, symmetric, non-negative) of class CsparseMatrix}

\item{k}{integer giving the number of mutual nearest meighbors
Defaults to the rounded square root of the number of variables.}

\item{quant}{Edges with weights > quantile(edge_weights, quant)  are retained in order to avoid large breaks}
}
\value{
a sparse Matrix of class CsparseMatrix with at most \code{length(S@x)} non-zero elements.
}
\description{
Keep k mutual nearest neighbors of variables given a sparse similarity matrix.  
By default 10 nearest neighbors of each node are retained to avoid node isolation
}
