% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{mapLabelsExhaustive}
\alias{mapLabelsExhaustive}
\title{Map class labels using an exhaustive algorithm.}
\usage{
mapLabelsExhaustive(a, b = NULL)
}
\arguments{
\item{a}{Either a numeric, charactor of factor with class labels, or if \code{b}
is \code{NULL}, a matrix or two-dimensional array (e.g. returned by \code{table})
corresponding to the confusion matrix.}

\item{b}{Either a numeric, charactor of factor with class labels of the same
length as \code{a}, or \code{NULL} if \code{a} is the confusion matrix.}
}
\value{
A character vector with the mapping of labels from \code{a} (names) to
    labels in \code{b} (values, coerced to \code{character}).
}
\description{
Map class labels (levels) of a given classification \code{a} in an exhaustive
fashion to its corresponding labels (levels) of a second classifciation \code{b}.
}
\details{
Each label in \code{a} is assigned to the label from \code{b} that minimizes
the classification error, exhaustively searching over all possible permutations
of label assignments.
}
\seealso{
\code{\link{mapLabelsGreedy}} for a greedy version that is
    faster especially for many classes. \code{\link{classError}} to calculate classification error.
}
