% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{mapLabelsGreedy}
\alias{mapLabelsGreedy}
\title{Map class labels using a greedy heuristic.}
\usage{
mapLabelsGreedy(a, b = NULL)
}
\arguments{
\item{a}{Either a numeric, charactor of factor with class labels, or if \code{b}
is \code{NULL}, a matrix or two-dimensional array (e.g. returned by \code{table})
corresponding to the confusion matrix.}

\item{b}{Either a numeric, charactor of factor with class labels of the same
length as \code{a}, or \code{NULL} if \code{a} is the confusion matrix.}
}
\value{
A character vector with the mapping of labels from \code{a} (names) to
    labels in \code{b} (values, coerced to \code{character}).
}
\description{
Map class labels (levels) of a given classification \code{a} in a greedy
fashion to its corresponding labels (levels) of a second classifciation \code{b}.
}
\details{
Each label in \code{a} is assigned to the label from \code{b} that it most
frequently co-occurs with. If there are multiple equally frequent pairs,
an arbitrary one of them is selected.
}
\seealso{
\code{\link{mapLabelsExhaustive}} for an exhaustive version that is
    guaranteed to return the optimal mapping that minimizes the classification
    error. \code{\link{classError}} to calculate classification error.
}
