% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{mycircle}
\alias{mycircle}
\title{Generate custom vertices for igraph plotting}
\usage{
mycircle(coords, v = NULL, params)
}
\arguments{
\item{coords}{The coordinates of the vertices, a matrix with two columns.}

\item{v}{The ids of the vertices to plot. It should match the number of rows in the coords argument.}

\item{params}{This is a function object that can be called to query graphical parameters.}
}
\value{
This function is used for its side effect of plotting graph vertices.
}
\description{
A plotting function for graph nodes used in graph visualization, passed to
\code{igraph::add.vertex.shape(..., plot=mycircle)}.
}
\seealso{
\code{\link[igraph]{add.vertex.shape}}.
}
