% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PCanberraMatOMP}
\alias{PCanberraMatOMP}
\title{Canberra distance between columns of two matrices (OpenMP version).}
\usage{
PCanberraMatOMP(A, B, nthreads = 2L)
}
\arguments{
\item{A}{Numeric matrix (variables by features)}

\item{B}{Numeric matrix (variables by features)}

\item{nthreads}{Integer specifying the number of OpenMP threads to use in
parallel parts (defaults to two, ignored on systems not supporting
OpenMP).}
}
\value{
A matrix of dimensions \code{ncol(A})-by-\code{ncol(B)}.
}
\description{
\code{PCanberraMatOMP} returns a matrix of Canberra distances between
the columns of the two matrices passed as arguments.
}
\details{
This function calculates the Canberra distances between the columns of two
matrices given as argument. It is implemented in C++ for efficiency. For two
argument matrices with the same number of r rows and c1 , c2 columns, the
return value is an c1-by-c2 matrix with the pairwise c1i c2j distances.
}
\examples{
x <- matrix(1:12, nrow=3, ncol=4)
y <- matrix(1:15, nrow=3, ncol=5)
PHellingerMatOMP(x, y)

}
