% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCellCycle.R
\name{predictCellCycle}
\alias{predictCellCycle}
\title{Predict (infer) cell cycle stage.}
\usage{
predictCellCycle(cnt, org = c("human", "mouse"), cor_thr = 0.3,
  granularity = c("high", "low"))
}
\arguments{
\item{cnt}{gene-by-cell count matrix (raw or normalized counts). Rownames must
be Entrez Gene identifiers.}

\item{org}{Character scalar specifying the species; one of "human"
(default) or "mouse".}

\item{cor_thr}{Numeric scalar value in [0,1] used to filter known cell cycle
genes (see Details).}

\item{granularity}{Character scaler selecting the level of detail for cell
cycle classification. \code{"high"} (default) will call 10 different cell
cycle stages, \code{"low"} will call only 5 different stages.}
}
\value{
A factor of length \code{ncol(cnt)} (the number of cells) with
    predicted cell cycle stages.
}
\description{
Predict (infert) cell cycle stage for individual cells using
    sets of known cell cycle-regulated genes.
}
\details{
\code{predictCellCycle} uses genes known to peak in transcription at
    given cell cycle stages (for human obtained from Whitfield et al., 2002, dowloaded
    from broadinstitute.org, see references; for mouse 1-to-1 orthologs of the
    human genes, as defined in Homologene release 68 were used).
    For each cell, five normalized cell cycle stage scores are calculated, and
    the cell is assigned to the one that most closely resembles the expected
    profile based on correlation.
}
\references{
Whitfield et al., "Identification of genes periodically expressed
    in the human cell cycle and their expression in tumors.", Mol Bio Cell, 2002,
    \url{http://www.ncbi.nlm.nih.gov/pubmed/12058064},
    G1_S genes: \url{http://software.broadinstitute.org/gsea/msigdb/geneset_page.jsp?geneSetName=WHITFIELD_CELL_CYCLE_G1_S},
    G2 genes: \url{http://software.broadinstitute.org/gsea/msigdb/geneset_page.jsp?geneSetName=WHITFIELD_CELL_CYCLE_G2},
    G2_M genes: \url{http://software.broadinstitute.org/gsea/msigdb/geneset_page.jsp?geneSetName=WHITFIELD_CELL_CYCLE_G2_M},
    M_G1 genes: \url{http://software.broadinstitute.org/gsea/msigdb/geneset_page.jsp?geneSetName=WHITFIELD_CELL_CYCLE_M_G1},
    S genes: \url{http://software.broadinstitute.org/gsea/msigdb/geneset_page.jsp?geneSetName=WHITFIELD_CELL_CYCLE_S}
}
