% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{RandString}
\alias{RandString}
\title{Generate a random alpha-numeric string}
\usage{
RandString(n = 1, len = 5)
}
\arguments{
\item{n}{numeric(1) specifying the number of strings to generate (default: 1)}

\item{len}{the number of characters per string (default: 5)}
}
\value{
a character vector with \code{n} random strings with \code{len} characters each.
}
\description{
Generate a random alpha-numeric string
}
\details{
Generate a character vector of \code{n} random strings of length \code{len}
consisting of alpha-numeric characters (0-9, a-z or A-Z), for example to use
as output files prefixes.
}
