% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC_cluster2.R
\name{SC_cluster2}
\alias{SC_cluster2}
\title{Unsupervised graph-based clustering of single cell RNA-seq data.}
\usage{
SC_cluster2(DM, use.par = FALSE, ncores = "all", is.cor = FALSE,
  impute = FALSE, filter = FALSE, rho = 0.25, pr.iter = 1,
  batch.penalty = 0.5, comm.method = igraph::cluster_infomap, ncom = NULL,
  ClassAssignment = rep(1, ncol(DM)), BatchAssignment = NULL,
  plotG = TRUE, maxG = 2500, fsuffix = RandString(),
  image.format = "png", ...)
}
\arguments{
\item{DM}{Count data (n genes-by-k cells) or directly a correlation k-by-k matrix.
Required argument.}

\item{use.par}{If \code{TRUE}, use parallel versions of distance calculation
functions based on \code{\link[foreach]{foreach}} (see details).}

\item{ncores}{a numeric(1) or character(1), either specifying the number of
parallel jobs to use, or \code{"all"} (default) to use up to 90 percent of
the available cores. Ignored if \code{use.par==FALSE}.}

\item{is.cor}{If \code{TRUE}, \code{DM} is assumed to be a correlation matrix.
Otherwise (the default), \code{DM} is assumed to be a genes-by-cells count
matrix, and a correlation matrix will be computed.}

\item{impute}{Perform imputation prior to clustering. Default: \code{FALSE}.}

\item{filter}{T/F Filter genes according to cv=f( mean ) fitting. Default: \code{TRUE}.}

\item{rho}{Inverse covariance matrix regularization (graph sparsification) parameter -> [0,1].
Default=0.25.  The parameter is then automatically scaled to account for
number of variables or converted into a matrix and adjusted according to
the batch.penalty factor to account for BatchAssignment (if given).}

\item{pr.iter}{A numberic scalar defining the number of iterations to use for
re-weighting the graph edges using personalized PageRank node similarity score.
Set to zero to deactivate.}

\item{batch.penalty}{[0,1] rho scaling factor for enforcing topological constraints
variables according to \code{BatchAssignment}. For penalty p  -> rho_same_batch=rho^(1-p),
rho_diff_batch=rho^(1+p). It is ignored if \code{BatchAssignment==NULL}.}

\item{comm.method}{Community detection algorithm. See igraph "communities"}

\item{ncom}{Forces the community detection algorithm to a fixed number of communities. Only possible with hierarchical methods (like fast_greedy).
If \code{NULL} (default) then the optimal determined number of clusters of the used  community detection algorithm.}

\item{ClassAssignment}{If available a numeric vector of length \code{k} with numeric
class labels (e.g-> c(1,2,1,1,1,2,3,3,3,1,2))}

\item{BatchAssignment}{If available a numeric vector of length \code{k} with numeric
batch labels (e.g-> c(1,2,1,1,1,2,3,3,3,1,2))}

\item{plotG}{if \code{TRUE} plots the resulting graph}

\item{maxG}{Approximate maximal number of vertices to include when plotting the graph.}

\item{fsuffix}{A suffix added to the file names of output plots. If not given
it will use a random 5 character string.}

\item{image.format}{Specifies the format of the created images. Currently only pdf and png filetypes are supported.}
}
\value{
Currently a list with the clustering results.
}
\description{
\code{SC_cluster} takes a gene-by-cell matrix with read counts as
    input and groups individual cells into clusters of similar cell (cell types).
    The number of cell types is automatically determined.
}
\details{
TODO
}
