% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{sparsify}
\alias{sparsify}
\title{Prune an affinity (weighted adjacency) matrix to keep only top \code{pct} fraction neighbors}
\usage{
sparsify(x, pct = 0.1)
}
\arguments{
\item{x}{weighted adjacency matrix represting the grpah to be pruned.}

\item{pct}{fraction of edges to keep.}
}
\value{
a matrix with the same dimensions as \code{x}, with only \code{pct} fraction of non-zero values.
}
\description{
Prune an affinity (weighted adjacency) matrix to keep only top \code{pct} fraction neighbors
}
