% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{sparsify2}
\alias{sparsify2}
\title{Prune an affinity (weighted adjacency) matrix to keep only top \code{pct} fraction neighbors}
\usage{
sparsify2(S, quant = 0.1)
}
\arguments{
\item{S}{a sparse similarity matrix (square, symmetric, non-negative) of class CsparseMatrix.}

\item{quant}{fraction of edges to keep.}
}
\value{
a sparse Matrix of class CsparseMatrix with a fraction of \code{pct} non-zero elements per column.
}
\description{
Prune an affinity (weighted adjacency) matrix to keep only top \code{pct} fraction neighbors
}
