% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPearsonMatOMP}
\alias{SPearsonMatOMP}
\title{Pearson's correlation coefficient between columns of a sparse matrix
    (OpenMP version).}
\usage{
SPearsonMatOMP(A, nthreads = 2L)
}
\arguments{
\item{A}{Numeric matrix (variables by features). Must be a \code{dgCMatrix}.}

\item{nthreads}{Integer specifying the number of OpenMP threads to use in
parallel parts (defaults to two, ignored on systems not supporting
OpenMP).}
}
\value{
A matrix of dimensions \code{ncol(A})-by-\code{ncol(B)}.
}
\description{
\code{SPearsonMatOMP} returns a symmetric matrix of Pearson's correlation
coefficients between all pairs of columns of the sparse matrix passed as argument.
}
\details{
This function calculates the Pearson's correlation coefficients between all
pairs of columns of the sparse matrix given as argument. It is implemented
in C++ for efficiency. For a matrix with c columns, the return
value is an c-by-c matrix with the pairwise ci cj correlation coefficients.
}
\examples{
x <- matrix(c(1,0,4,0,0,1,2,5,0,3,0,0,1,3,0), nrow=5)
cor(x)
xs <- as(x, "sparseMatrix")
SPearsonMatOMP(xs)

}
