% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ssPCanberraMatOMP}
\alias{ssPCanberraMatOMP}
\title{sparse-sparse verion of PCanberraMatOMP}
\usage{
ssPCanberraMatOMP(A, B, nthreads = 2L)
}
\arguments{
\item{A}{Sparse numeric matrix (variables by features). Must be a \code{dgCMatrix}.}

\item{B}{Sparse numeric matrix (variables by features). Must be a \code{dgCMatrix}.}

\item{nthreads}{Integer specifying the number of OpenMP threads to use in
parallel parts (defaults to two, ignored on systems not supporting
OpenMP).}
}
\value{
A matrix of dimensions \code{ncol(A})-by-\code{ncol(B)}.
}
\description{
\code{ssPCanberraMatOMP} is an alternative version of \code{PCanberraMatOMP}
for sparse inputs.
}
\examples{
x <- matrix(c(1,0,4,0,0,1,2,5,0,3,0,0,1,3,0), nrow=5)
xs <- as(x, "sparseMatrix")

PCanberraMatOMP(x, x)
ssPCanberraMatOMP(xs, xs)

}
