% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC_cluster.R
\name{WScor}
\alias{WScor}
\title{Combined distance calculations.}
\usage{
WScor(M, C1 = matrix(1, ncol(M), ncol(M)), CanberraDist = CanberraDist,
  SpearmanCor = SpearmanCor, HellingerDist = HellingerDist,
  ShrinkCor = ShrinkCor)
}
\arguments{
\item{M}{gene-by-cell count matrix.}

\item{C1}{cell-by-cell (correlation?) matrix.}

\item{CanberraDist}{Function to calculate Canberra distance.}

\item{SpearmanCor}{Function to calculate Spearman rank correlation.}

\item{HellingerDist}{Function to calculate Hellinger distance.}

\item{ShrinkCor}{Function to calculate shrinkage correlation.}
}
\value{
cell-by-cell distance matrix.
}
\description{
Calculate and combine several between-cell distance measures to
    obtain robust cell-to-cell distances.
}
