% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC_cluster2.R
\name{WScor2}
\alias{WScor2}
\title{Combined distance calculations.}
\usage{
WScor2(M, PearsonCor = PearsonCor, ShrinkCor = ShrinkCor)
}
\arguments{
\item{M}{gene-by-cell count matrix.}

\item{ShrinkCor}{Function to calculate shrinkage correlation.}

\item{C1}{cell-by-cell (correlation?) matrix.}

\item{CanberraDist}{Function to calculate Canberra distance.}

\item{SpearmanCor}{Function to calculate Spearman rank correlation.}

\item{HellingerDist}{Function to calculate Hellinger distance.}
}
\value{
cell-by-cell distance matrix.
}
\description{
Calculate and combine several between-cell distance measures to
    obtain robust cell-to-cell distances.
}
