% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griph_cluster.R
\name{WScorFB}
\alias{WScorFB}
\title{Combined distance calculations.}
\usage{
WScorFB(M, FB, ShrinkCor = ShrinkCor)
}
\arguments{
\item{M}{gene-by-cell count matrix.}

\item{FB}{gene-by-FakeBulks gene sums count matrix.}

\item{ShrinkCor}{Function to calculate shrinkage correlation.}
}
\value{
cell-by-cell distance matrix.
}
\description{
Calculate and combine several between-cell distance measures to
    obtain robust cell-to-cell distances.
}
