% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{get.knn}
\alias{get.knn}
\title{Get k nearest neighbors of variables given a feature matrix.}
\usage{
get.knn(S, k = round(sqrt(nrow(S))))
}
\arguments{
\item{S}{similarity matrix (here: always square)}

\item{k}{integer giving the number of nearest
neighbors to be returned. Defaults to the
rounded square root of the number of rows (columns).}
}
\value{
A \code{k} by \code{ncol(S)} matrix.
}
\description{
Get k nearest neighbors of variables given a feature matrix.
}
