% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griph_cluster.R
\name{griph_cluster}
\alias{griph_cluster}
\title{Wrapper function for SC_cluster}
\usage{
griph_cluster(DM, K = NULL, SamplingSize = NULL, ref.iter = 1,
  use.par = TRUE, ncores = "all", filter = TRUE, rho = 0.25,
  batch.penalty = 0.5, seed = 127350, ClassAssignment = rep(1, ncol(DM)),
  BatchAssignment = NULL, ncom = NULL, plot_ = TRUE, maxG = 2500,
  fsuffix = NULL, image.format = "png",
  comm.method = igraph::cluster_louvain)
}
\arguments{
\item{DM}{Count data (n genes-by-k cells) or directly a correlation k-by-k matrix.
Required argument.}

\item{K}{Number of nearest neighbors to consider. If \code{NULL} (default) it is set to to \code{0.25*sqrt(k)+4*D}, D being
the dataset dimensionality determined in the initialization (sampling) step. Smaller K values result in more disconnected
cell clusters and vice versa. Typically should be in the range [20,200].}

\item{SamplingSize}{Number of sampled cells in initialization step. If NULL it is set to \code{min(2000,k/2)}}

\item{ref.iter}{Number of clustering refinement iterations.  If set to 0 only the clustering initialization 
step is performed to \code{min(SamplingSize,ncol(DM))} cells.}

\item{use.par}{If \code{TRUE}, use parallel versions of distance calculation
functions based on \code{\link[foreach]{foreach}} (see details).}

\item{ncores}{a numeric(1) or character(1), either specifying the number of
parallel jobs to use, or \code{"all"} (default) to use up to 90 percent of
the available cores. Ignored if \code{use.par==FALSE}.}

\item{filter}{TRUE/FALSE or a positive numeric values. Determines  whether filtering for retaining only overdispersed genes should be performed.
If FALSE no gene filtering is performed. If TRUE (default) only the top 25% overdispersed genes are retained.
If set to a number between 0 and 1 than this fraction of genes is retained.
If set to an integer > 1 than this number of genes is retained}

\item{rho}{Inverse covariance matrix regularization (graph sparsification) parameter -> [0,1].
Default=0.25.  The parameter is then automatically scaled to account for
number of variables or converted into a matrix and adjusted according to
the batch.penalty factor to account for BatchAssignment (if given).}

\item{batch.penalty}{[0,1] rho scaling factor for enforcing topological constraints
variables according to \code{BatchAssignment}. For penalty p  -> rho_same_batch=rho^(1-p),
rho_diff_batch=rho^(1+p). It is ignored if \code{BatchAssignment==NULL}.}

\item{seed}{Set seed for reproducible results.}

\item{ClassAssignment}{If available a numeric vector of length \code{k} with numeric
class labels (e.g-> c(1,2,1,1,1,2,3,3,3,1,2))}

\item{BatchAssignment}{If available a numeric vector of length \code{k} with numeric
batch labels (e.g-> c(1,2,1,1,1,2,3,3,3,1,2))}

\item{ncom}{Forces detection of a fixed number of clusters (communities).}

\item{plot_}{if \code{TRUE} plots the obtained graph using \code{plotLVis}. The resulting
2D-embedding is stored in the \code{plotLVis} component of the result.}

\item{maxG}{Approximate maximal number of vertices to include when plotting the graph.}

\item{fsuffix}{A suffix added to the file names of output plots. If \code{NULL} (default),
it will use a random 5 character string.}

\item{image.format}{Specifies the format of the created images. Currently only pdf and png filetypes are supported.}

\item{comm.method}{Community detection algorithm. See igraph "communities". By default multilevel louvain is used.}
}
\value{
Currently a list with the clustering results:

MEMB: A vector of membership assignment for each cell

MEMB.true: A vector with the true cell assignment if one was specified.

DISTM: A sparse (dgCMatrix) k x k cell distance matrix

ConfMatrix: Confusion matrix based on the MEMB and MEMB.true vectors.

miscl: Misclassification error based on the MEMB and MEMB.true vectors.

GRAO: An igraph graph object modelling the cell population

plotLVis: The plotLVis projection of the graph if plot_ was set to try. NULL otherwise

complete_MEMB: The igraph communities object returned by the community detection algorithm

GeneList: A character vector with the genes that survived the filtering.
}
\description{
\code{griph_cluster} takes a gene-by-cell matrix with read counts as
    input and groups individual cells into clusters of similar cell (cell types).
    The number of cell types is automatically determined.
}
\details{
TODO
}
