% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotGraph}
\alias{plotGraph}
\title{Visualize griph result as a graph.}
\usage{
plotGraph(gr, maxG = 2500, fill.type = "predicted", line.type = "true",
  mark.type = "none", collapse.type = c("none", "predicted", "true"),
  fill.col = "Spectral", line.col = "Dark2", mark.col = "Pastel1",
  draw.edges = NULL, seed = 91919, fsuffix = NULL, image.format = NA,
  forceRecalculation = FALSE, quiet = FALSE, plot.args = list(pch = 21L,
  cex = 1.5, lwd = 2.5))
}
\arguments{
\item{gr}{A \code{griph} result, as returned by \code{\link{SC_cluster}} or
\code{\link{griph_cluster}}.}

\item{maxG}{Approximate maximal number of vertices to include when plotting
the graph.}

\item{fill.type}{Type of fill color, one of \code{"predicted"} (predicted class
labels, default), \code{"true"} (true class labels, if available), \code{"none"}
(no fill color), a numeric vectors  or a factor whose values should be
mapped to fill colors, using the palette given in \code{fill.col}.}

\item{line.type}{Type of line color, as in \code{fill.type}.}

\item{mark.type}{Type of cell class defnition to mark using polygons, as in \code{fill.type}.}

\item{collapse.type}{Type of cell class to use for graph simplification,
by combining cells of the same class into a single vertex. If set to a value
other than \code{"none"}, the same value will also be used for \code{fill.type}
and \code{line.type} is ignored.}

\item{fill.col}{Fill color palette, either a character scalar with a \code{\link{RColorBrewer}}
color palette name or color vector defining the palette to use for vertex fill coloring.}

\item{line.col}{Line color palette, as in \code{fill.col}.}

\item{mark.col}{Mark color palette, as in \code{fill.col}.}

\item{draw.edges}{If \code{NULL} (default), draw edges if \code{collapse.type != "none"}.
\code{TRUE} or \code{FALSE} can be used to override the default.}

\item{seed}{Random number seed to make graph layout deterministic.}

\item{fsuffix}{A suffix added to the file names of output plots. If \code{NULL} (default),
it will use a random 5 character string. Ignored if \code{image.format} is \code{NULL}.}

\item{image.format}{Specifies the format of the created image. Currently supported are
\code{\link{pdf}}, \code{\link{png}} or \code{NA}. If \code{NA} (the default), the plot
is rendered on the currently opened plotting device.}

\item{forceRecalculation}{If \code{TRUE}, recalculate plotting-optimized graph
even if it is already contained in \code{gr}.}

\item{quiet}{If \code{TRUE}, do not report on progress.}

\item{plot.args}{a named list of  arguments for controlling specific plot points attributes (cex, lwd, pch -should between 21 and 25- )}
}
\value{
Invisibly the plot-optimized version of the graph as an \code{igraph} object.
}
\description{
Plot a graph obtained from \code{\link{SC_cluster}} or
    \code{\link{griph_cluster}}, allowing to apply graph subsampling and
    control coloring.
}
