% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{psparse.cor}
\alias{psparse.cor}
\title{Pairwise correlation for 2 sparse matrices (with the same number of rows)}
\usage{
psparse.cor(x, y)
}
\arguments{
\item{x}{a sparse matrix.}

\item{y}{a sparse matrix (same number of rows as \code{x}).}
}
\value{
a matrix with correlations between columns of the inputs \code{x} and \code{y}.
}
\description{
Pairwise correlation for 2 sparse matrices (with the same number of rows)
}
