% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC_cluster.R
\name{WScor}
\alias{WScor}
\title{Combined distance calculations.}
\usage{
WScor(M, PearsonCor = PearsonCor, ShrinkCor = ShrinkCor)
}
\arguments{
\item{M}{gene-by-cell count matrix.}

\item{PearsonCor}{Function to calculate Pearson correlation.}

\item{ShrinkCor}{Function to calculate shrinkage correlation.}
}
\value{
cell-by-cell distance matrix.
}
\description{
Calculate and combine several between-cell distance measures to
    obtain robust cell-to-cell distances.
}
