% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC_cluster.R
\name{WScor}
\alias{WScor}
\title{Combined distance calculations.}
\usage{
WScor(M, PPearsonCor, PSpearmanCor, PHellinger, PCanberra,
  ShrinkCor = ShrinkCor, seed = 127350)
}
\arguments{
\item{M}{gene-by-cell count matrix.}

\item{PPearsonCor}{Function to calculate Pearson correlation between the columns of two matrices.}

\item{PSpearmanCor}{Function to calculate Spearman correlation between the columns of two matrices.}

\item{PHellinger}{Function to calculate Hellinger Distance between the columns of two matrices.}

\item{PCanberra}{Function to calculate Canberra Distance between the columns of two matrices.}

\item{ShrinkCor}{Function to calculate shrinkage correlation.}

\item{seed}{Set seed for reproducible results.}
}
\value{
cell-by-cell distance matrix.
}
\description{
Calculate and combine several between-cell distance measures to
    obtain robust cell-to-cell distances.
}
