% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griph_cluster.R
\name{WScorFB}
\alias{WScorFB}
\title{Combined distance calculations.}
\usage{
WScorFB(M, FB, K = 50, PSpearmanCor, PPearsonCor, PHellinger, PCanberra,
  ShrinkCor = ShrinkCor)
}
\arguments{
\item{M}{gene-by-cell count matrix.}

\item{FB}{gene-by-FakeBulks gene sums count matrix.}

\item{K}{Number of nearest neighbors to retain}

\item{PSpearmanCor}{Function to calculate Spearman correlation between the columns of two matrices.}

\item{PPearsonCor}{Function to calculate Pearson correlation between the columns of two matrices.}

\item{PHellinger}{Function to calculate Hellinger Distance between the columns of two matrices.}

\item{PCanberra}{Function to calculate Canberra Distance between the columns of two matrices.}

\item{ShrinkCor}{Function to calculate shrinkage correlation.}
}
\value{
cell-by-cell distance matrix.
}
\description{
Calculate and combine several between-cell distance measures to
    obtain robust cell-to-cell distances.
}
